# Keep a `ByteData` around in `CodedBufferReader` to avoid repeated `ByteData` allocs